Refactor to use the dataset description model from the HABS converter
