+++ clinica/iotools/converters/oasis3_to_bids/oasis3_utils.py
@@ -322,7 +322,7 @@ def write_bids(
                 write_to_tsv(sessions_group, sessions_file)
 
         with fs.open(to / "dataset_description.json", mode="w") as f:
-            BidsDatasetDescription(name="OASIS-3", bids_version="1.6.0").write(to=f)
+            BidsDatasetDescription(name="OASIS-3").save(to=f)
 
     # Perform import of imaging data next.
     for filename, metadata in scans.iterrows():
